wildcard: remove files and move functions into ftplistparser.c
